Unify the object-creation expressions we produce.
