+++ src/EditorFeatures/CSharpTest/UseObjectInitializer/UseObjectInitializerTests.cs
@@ -36,7 +36,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
 
     void M()
     {
-        var c = new C()
+        var c = new C
         {
             i = 1
         };
@@ -65,7 +65,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
 
     void M()
     {
-        var c = new C()
+        var c = new C
         {
             i = 1
         };
@@ -113,7 +113,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
     void M()
     {
         C c;
-        c = new C()
+        c = new C
         {
             i = 1
         };
@@ -161,7 +161,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
     void M()
     {
         C c = null;
-        c = new C()
+        c = new C
         {
             i = 1
         };
@@ -190,7 +190,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
 
     void M()
     {
-        var c = new C()
+        var c = new C
         {
             i = 1
         };
@@ -224,7 +224,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
     void M()
     {
         C[] array;
-        array[0] = new C()
+        array[0] = new C
         {
             i = 1,
             j = 2
@@ -256,7 +256,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
 
     void M()
     {
-        var c = new C()
+        var c = new C
         {
             i = 1
         };
@@ -325,7 +325,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
     void M()
     {
         var v = new C(() => {
-            var v2 = new C()
+            var v2 = new C
             {
                 i = 1
             };
@@ -362,10 +362,10 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
 
     void M()
     {
-        var v = new C()
+        var v = new C
         {
             j = () => {
-                var v2 = new C()
+                var v2 = new C
                 {
                     i = 1
                 };
@@ -403,12 +403,12 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.UseObjectInitializer
     void M()
     {
         C[] array;
-        array[0] = new C()
+        array[0] = new C
         {
             i = 1,
             j = 2
         };
-        array[1] = new C()
+        array[1] = new C
         {
             i = 3,
             j = 4
@@ -440,7 +440,7 @@ class C
     int j;
     void M()
     {
-        var c = new C()
+        var c = new C
         {
             i = 1, // Foo
             j = 2 // Bar
