+++ src/Features/CSharp/Portable/UseObjectInitializer/CSharpUseObjectInitializerCodeFixProvider.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Composition;
@@ -34,13 +33,8 @@ namespace Microsoft.CodeAnalysis.CSharp.UseObjectInitializer
             ObjectCreationExpressionSyntax objectCreation,
             ImmutableArray<Match<ExpressionSyntax, StatementSyntax, MemberAccessExpressionSyntax, ExpressionStatementSyntax>> matches)
         {
-            var openBrace = SyntaxFactory.Token(SyntaxKind.OpenBraceToken)
-                                         .WithTrailingTrivia(SyntaxFactory.ElasticCarriageReturnLineFeed);
-            var initializer = SyntaxFactory.InitializerExpression(
-                SyntaxKind.ObjectInitializerExpression,
-                CreateExpressions(matches)).WithOpenBraceToken(openBrace);
-
-            return objectCreation.WithInitializer(initializer);
+            return UseInitializerHelpers.GetNewObjectCreation(
+                objectCreation, CreateExpressions(matches));
         }
 
         private SeparatedSyntaxList<ExpressionSyntax> CreateExpressions(
