+++ src/Features/VisualBasic/Portable/UseObjectInitializer/VisualBasicUseObjectInitializerCodeFixProvider.vb
@@ -45,12 +45,10 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.UseObjectInitializer
                 objectCreation As ObjectCreationExpressionSyntax,
                 matches As ImmutableArray(Of Match(Of ExpressionSyntax, StatementSyntax, MemberAccessExpressionSyntax, AssignmentStatementSyntax))) As ObjectCreationExpressionSyntax
 
-            Dim initializer = SyntaxFactory.ObjectMemberInitializer(
-                CreateFieldInitializers(matches))
-
-            Return objectCreation.WithoutTrailingTrivia().
-                                  WithInitializer(initializer).
-                                  WithTrailingTrivia(objectCreation.GetTrailingTrivia())
+            Return UseInitializerHelpers.GetNewObjectCreation(
+                objectCreation,
+                SyntaxFactory.ObjectMemberInitializer(
+                    CreateFieldInitializers(matches)))
         End Function
 
         Private Function CreateFieldInitializers(
